rename management to livestock
