+++ apps/livestock/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class LivestockConfig(AppConfig):
+    name = "apps.livestock"
