+++ apps/livestock/views.py
