--- apps/management/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ManagementConfig(AppConfig):
-    name = "apps.management"
